main: use stdlib flags
